+++ src/Persistence/src/Dao/SQLite/SQLiteFlightDao.cpp
@@ -95,7 +95,7 @@ bool SQLiteFlightDao::exportFlightData(const FlightData &flightData) const noexc
 
 bool SQLiteFlightDao::get(std::int64_t id, FlightData &flightData) const noexcept
 {
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.setForwardOnly(true);
     query.prepare(
@@ -185,7 +185,7 @@ bool SQLiteFlightDao::get(std::int64_t id, FlightData &flightData) const noexcep
 
 bool SQLiteFlightDao::deleteById(std::int64_t id) const noexcept
 {
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.prepare(
         "delete "
@@ -208,7 +208,7 @@ bool SQLiteFlightDao::deleteById(std::int64_t id) const noexcept
 
 bool SQLiteFlightDao::updateTitle(std::int64_t id, const QString &title) const noexcept
 {
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.prepare(
         "update flight "
@@ -229,7 +229,7 @@ bool SQLiteFlightDao::updateTitle(std::int64_t id, const QString &title) const n
 
 bool SQLiteFlightDao::updateFlightNumber(std::int64_t id, const QString &flightNumber) const noexcept
 {
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.prepare(
         "update flight "
@@ -250,7 +250,7 @@ bool SQLiteFlightDao::updateFlightNumber(std::int64_t id, const QString &flightN
 
 bool SQLiteFlightDao::updateDescription(std::int64_t id, const QString &description) const noexcept
 {
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.prepare(
         "update flight "
@@ -271,7 +271,7 @@ bool SQLiteFlightDao::updateDescription(std::int64_t id, const QString &descript
 
 bool SQLiteFlightDao::updateUserAircraftIndex(std::int64_t id, int index) const noexcept
 {
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.prepare(
         "update flight "
@@ -295,7 +295,7 @@ inline std::int64_t SQLiteFlightDao::insertFlight(const FlightData &flightData) 
 {
     std::int64_t flightId {Const::InvalidId};
 
-    const QSqlDatabase db {QSqlDatabase::database(d->connectionName)};
+    const auto db {QSqlDatabase::database(d->connectionName)};
     QSqlQuery query {db};
     query.prepare(
         "insert into flight ("
@@ -351,7 +351,7 @@ inline std::int64_t SQLiteFlightDao::insertFlight(const FlightData &flightData) 
         ");"
     );
 
-    const FlightCondition &flightCondition = flightData.flightCondition;
+    const auto &flightCondition = flightData.flightCondition;
     query.bindValue(":creation_time", flightData.creationTime.toUTC());
     // Sequence number starts at 1
     query.bindValue(":user_aircraft_seq_nr", flightData.userAircraftIndex + 1);