+++ src/Plugins/Flight/Export/GpxExport/src/GpxExportPlugin.cpp
@@ -52,6 +52,11 @@
 
 struct GpxExportPluginPrivate
 {
+    GpxExportPluginPrivate()
+    {
+        startDateTimeUtc.setTimeZone(QTimeZone::utc());
+    }
+
     GpxExportSettings pluginSettings;
     QDateTime startDateTimeUtc;
     Unit unit;
@@ -111,7 +116,6 @@ bool GpxExportPlugin::exportFlightData(const FlightData &flightData, QIODevice &
 
 bool GpxExportPlugin::exportAircraft(const FlightData &flightData, const Aircraft &aircraft, QIODevice &io) const noexcept
 {
-    updateStartDateTimeUtc(flightData, aircraft);
     io.setTextModeEnabled(true);
     bool ok = exportHeader(io);
     if (ok) {
@@ -121,7 +125,7 @@ bool GpxExportPlugin::exportAircraft(const FlightData &flightData, const Aircraf
         ok = exportWaypoints(flightData, io);
     }
     if (ok) {
-        ok = exportSingleAircraft(aircraft, io);
+        ok = exportSingleAircraft(flightData, aircraft, io);
     }    
     if (ok) {
         ok = exportFooter(io);
@@ -170,7 +174,7 @@ bool GpxExportPlugin::exportAllAircraft(const FlightData &flightData, QIODevice 
 {
     bool ok {true};
     for (const auto &aircraft : flightData.aircraft) {
-        ok = exportSingleAircraft(aircraft, io);
+        ok = exportSingleAircraft(flightData, aircraft, io);
         if (!ok) {
             break;
         }
@@ -178,13 +182,14 @@ bool GpxExportPlugin::exportAllAircraft(const FlightData &flightData, QIODevice 
     return ok;
 }
 
-bool GpxExportPlugin::exportSingleAircraft(const Aircraft &aircraft, QIODevice &io) const noexcept
+bool GpxExportPlugin::exportSingleAircraft(const FlightData &flightData, const Aircraft &aircraft, QIODevice &io) const noexcept
 {
-    const std::vector<PositionData> interpolatedPositionData = Export::resamplePositionDataForExport(aircraft, d->pluginSettings.getResamplingPeriod());
+    updateStartDateTimeUtc(flightData, aircraft);
+    const auto interpolatedPositionData = Export::resamplePositionDataForExport(aircraft, d->pluginSettings.getResamplingPeriod());
     bool ok {true};
     if (interpolatedPositionData.size() > 0) {
 
-        const AircraftInfo &aircraftInfo = aircraft.getAircraftInfo();
+        const auto &aircraftInfo = aircraft.getAircraftInfo();
         const QString trackBegin =
 "  <trk>\n"
 "    <name><![CDATA[" % aircraftInfo.aircraftType.type % "]]></name>\n"
@@ -215,7 +220,7 @@ bool GpxExportPlugin::exportSingleAircraft(const Aircraft &aircraft, QIODevice &
 bool GpxExportPlugin::exportWaypoints(const FlightData &flightData, QIODevice &io) const noexcept
 {
     bool ok {true};
-    const FlightPlan &flightPlan = flightData.getUserAircraftConst().getFlightPlan();
+    const auto &flightPlan = flightData.getUserAircraftConst().getFlightPlan();
     for (const auto &waypoint : flightPlan) {
         ok = exportWaypoint(waypoint, io);
         if (!ok) {
@@ -233,7 +238,7 @@ bool GpxExportPlugin::exportFooter(QIODevice &io) const noexcept
 
 QString GpxExportPlugin::getFlightDescription(const FlightData &flightData) const noexcept
 {
-    const FlightCondition &flightCondition = flightData.flightCondition;
+    const auto &flightCondition = flightData.flightCondition;
     return flightData.description % "\n" %
            "\n" %
            QObject::tr("Creation date") % ": " % d->unit.formatDate(flightData.creationTime) % "\n" %
@@ -251,12 +256,12 @@ QString GpxExportPlugin::getFlightDescription(const FlightData &flightData) cons
 
 QString GpxExportPlugin::getAircraftDescription(const Aircraft &aircraft) const noexcept
 {
-    const AircraftInfo &info = aircraft.getAircraftInfo();
-    const AircraftType &type = info.aircraftType;
-    return QObject::tr("Category") % ": " % type.category % "\n" %
-           QObject::tr("Engine type") % ": " % SimType::engineTypeToString(type.engineType) % "\n" %
-           QObject::tr("Number of engines") % ": " % d->unit.formatNumber(type.numberOfEngines, 0) % "\n" %
-           QObject::tr("Wingspan") % ": " % d->unit.formatFeet(type.wingSpan) % "\n"
+    const auto &info = aircraft.getAircraftInfo();
+    const auto &aircraftType = info.aircraftType;
+    return QObject::tr("Category") % ": " % aircraftType.category % "\n" %
+           QObject::tr("Engine type") % ": " % SimType::engineTypeToString(aircraftType.engineType) % "\n" %
+           QObject::tr("Number of engines") % ": " % d->unit.formatNumber(aircraftType.numberOfEngines, 0) % "\n" %
+           QObject::tr("Wingspan") % ": " % d->unit.formatFeet(aircraftType.wingSpan) % "\n"
            "\n" %
            QObject::tr("Initial altitude above ground") % ": " % d->unit.formatFeet(info.altitudeAboveGround) % "\n" %
            QObject::tr("Initial airspeed") % ": " % d->unit.formatKnots(info.initialAirspeed) % "\n" %
@@ -266,27 +271,40 @@ QString GpxExportPlugin::getAircraftDescription(const Aircraft &aircraft) const 
 
 inline bool GpxExportPlugin::exportTrackPoint(const PositionData &positionData, QIODevice &io) const noexcept
 {
-    const QDateTime dateTimeUtc = d->startDateTimeUtc.addMSecs(positionData.timestamp);
-    // Convert height above EGM geoid to height above WGS84 ellipsoid (HAE) [meters]
-    const double heightAboveEllipsoid = d->convert.egmToWgs84Ellipsoid(Convert::feetToMeters(positionData.altitude), positionData.latitude, positionData.longitude);
-
-    const QString trackPoint =
+    // Elevation above mean sea level (MSL)
+    const auto elevation = Convert::feetToMeters(positionData.altitude);
+    const auto dateTimeUtc = d->startDateTimeUtc.addMSecs(positionData.timestamp);
+    QString trackPoint =
 "      <trkpt lat=\"" % Export::formatCoordinate(positionData.latitude) % "\" lon=\"" % Export::formatCoordinate(positionData.longitude) % "\">\n"
-"        <ele>" % Export::formatNumber(heightAboveEllipsoid).toUtf8() % "</ele>\n"
-"        <time>" % dateTimeUtc.toString(Qt::ISODate) % "</time>\n"
+"        <ele>" % Export::formatNumber(elevation).toUtf8() % "</ele>\n"
+"        <time>" % dateTimeUtc.toString(Qt::ISODateWithMs) % "</time>\n";
+    if (d->pluginSettings.isGeoidHeightExportEnabled()) {
+        // Calculate the geoid height
+        const auto geoidHeight = d->convert.geoidToEllipsoidHeight(0, positionData.latitude, positionData.longitude);
+        trackPoint = trackPoint %
+"        <geoidheight>" % Export::formatNumber(geoidHeight).toUtf8() % "</geoidheight>\n";
+    }
+    trackPoint = trackPoint %
 "      </trkpt>\n";
 
     return io.write(trackPoint.toUtf8());
 }
 
-inline bool GpxExportPlugin::exportWaypoint(const Waypoint &waypoint, QIODevice &io) const noexcept
+inline bool GpxExportPlugin::exportWaypoint(const auto &waypoint, QIODevice &io) const noexcept
 {
-    // Convert height above EGM geoid to height above WGS84 ellipsoid (HAE) [meters]
-    const double heightAboveEllipsoid = d->convert.egmToWgs84Ellipsoid(Convert::feetToMeters(waypoint.altitude), waypoint.latitude, waypoint.longitude);
-    const QString waypointString =
+    // Elevation above mean sea level (MSL)
+    const auto elevation = Convert::feetToMeters(waypoint.altitude);
+    QString waypointString =
 "  <wpt lat=\"" % Export::formatCoordinate(waypoint.latitude) % "\" lon=\"" % Export::formatCoordinate(waypoint.longitude) % "\">\n"
-"    <ele>" % Export::formatNumber(heightAboveEllipsoid).toUtf8() % "</ele>\n"
-"    <time>" % waypoint.zuluTime.toString(Qt::ISODate) % "</time>\n"
+"    <ele>" % Export::formatNumber(elevation).toUtf8() % "</ele>\n"
+"    <time>" % waypoint.zuluTime.toString(Qt::ISODateWithMs) % "</time>\n";
+    if (d->pluginSettings.isGeoidHeightExportEnabled()) {
+        // Calculate the geoid height
+        const auto geoidHeight = d->convert.geoidToEllipsoidHeight(0, waypoint.latitude, waypoint.longitude);
+        waypointString = waypointString %
+"    <geoidheight>" % Export::formatNumber(geoidHeight).toUtf8() % "</geoidheight>\n";
+    }
+    waypointString = waypointString %
 "    <name>" % waypoint.identifier % "</name>\n"
 "  </wpt>\n";