+++ src/Plugins/Flight/Import/GpxImport/src/GpxParser.cpp
@@ -376,7 +376,7 @@ bool GpxParser::parseWaypointType(double &latitude, double &longitude, double &a
 
     if (d->pluginSettings.isConvertAltitudeEnabled()) {
         // Convert height above WGS84 ellipsoid (HAE) to height above EGM geoid [meters]
-        altitude = d->convert.wgs84ToEgmGeoid(altitude, latitude, longitude);
+        altitude = d->convert.ellipsoidToGeoidHeight(altitude, latitude, longitude);
     }
     altitude = Convert::metersToFeet(altitude);